Metrics moved to a separate module metrics.py
